requirements.txt
kafka-python==2.0.2
faker==24.8.0
python-dotenv==1.0.1
pandas==2.0.3